Keywords: Database Systems, PHP, Web Application,
Web application that allows user to to post their complaints related to
public service departments like Police, Fire etc. These Complaints will
be posted as tweets from a common portal to twitter wherein other people
facing the same problem can either comment/re-tweet or like/favorite the
tweet. These tweets/complaints will be extracted from the account on
twitter on a predefined time interval and posted to the respective
department dashboard sorted/prioritized based on the highest number of
re-tweets or likes. In this way, we will prioritize the most common
complaint/problem faced by people and ensure that the most important or
most common issues are resolved faster. On resolution of the
issue/complaint by the respective department, the user will be notified
about the same.